refactor: Move util functions into folder
